refactor : 계층형 패키지 스타일로 수정
