+++ app/src/main/java/com/codesoom/assignment/product/application/ProductDeleter.java
@@ -1,0 +1,20 @@
+package com.codesoom.assignment.product.application;
+
+import com.codesoom.assignment.product.domain.Product;
+import com.codesoom.assignment.product.infra.persistence.ProductRepository;
+import org.springframework.stereotype.Service;
+
+@Service
+public class ProductDeleter {
+
+    private final ProductRepository productRepository;
+
+    public ProductDeleter(ProductRepository productRepository) {
+        this.productRepository = productRepository;
+    }
+
+    public void deleteProduct(Long id) {
+        Product product = productRepository.findById(id).orElseThrow(ProductNotFoundException::new);
+        productRepository.delete(product);
+    }
+}
