+++ app/src/main/java/com/codesoom/assignment/product/application/ProductReader.java
@@ -1,0 +1,23 @@
+package com.codesoom.assignment.product.application;
+
+import com.codesoom.assignment.product.domain.Product;
+import com.codesoom.assignment.product.infra.persistence.ProductRepository;
+import org.springframework.stereotype.Service;
+
+@Service
+public class ProductReader {
+
+    private final ProductRepository productRepository;
+
+    public ProductReader(ProductRepository productRepository) {
+        this.productRepository = productRepository;
+    }
+
+    public Iterable<Product> getProductList() {
+        return productRepository.findAll();
+    }
+
+    public Product getProduct(Long id) {
+        return productRepository.findById(id).orElseThrow(ProductNotFoundException::new);
+    }
+}
