+++ app/src/main/java/com/codesoom/assignment/product/application/ProductUpdater.java
@@ -3,30 +3,16 @@ package com.codesoom.assignment.product.application;
 import com.codesoom.assignment.product.domain.Product;
 import com.codesoom.assignment.product.domain.dto.ProductRequest;
 import com.codesoom.assignment.product.infra.persistence.ProductRepository;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
 
-@Slf4j
 @Service
-public class ProductService {
+public class ProductUpdater {
     private final ProductRepository productRepository;
 
-    public ProductService(ProductRepository productRepository) {
+    public ProductUpdater(ProductRepository productRepository) {
         this.productRepository = productRepository;
-    }
-
-    public Product createProduct(ProductRequest product) {
-        return productRepository.save(product.toProductEntity());
-    }
-
-    public Iterable<Product> getProductList() {
-        return productRepository.findAll();
-    }
-
-    public Product getProduct(Long id) {
-        return productRepository.findById(id).orElseThrow(ProductNotFoundException::new);
     }
 
     @Transactional
@@ -34,10 +20,5 @@ public class ProductService {
         Product product = productRepository.findById(id).orElseThrow(ProductNotFoundException::new);
         product.update(productRequest.toProductEntity());
         return product;
-    }
-
-    public void deleteProduct(Long id) {
-        Product product = productRepository.findById(id).orElseThrow(ProductNotFoundException::new);
-        productRepository.delete(product);
     }
 }
